requirements/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
arrow==1.1.0
asgiref==3.3.4
astroid==2.5.6
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==21.4b1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
channels==3.0.3
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cookiecutter==1.7.2
coverage==5.5
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.8
django-allauth==0.44.0
django-cors-headers==3.7.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-grappelli==2.15.1
django-jazzmin==2.4.6
django-markdownify==0.9.0
django-model-utils==4.1.1
django-redis==4.12.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
docutils==0.16
factory-boy==3.2.0
Faker==8.1.1
filelock==3.0.12
filetype==1.0.7
flake8==3.9.1
flake8-isort==4.0.0
hiredis==1.1.0
hyperlink==21.0.0
identify==2.2.4
idna==2.10
imagesize==1.2.0
incremental==21.3.0
iniconfig==1.1.1
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
jsonlines==2.0.0
lazy-object-proxy==1.6.0
livereload==2.6.3
Markdown==3.3.4
markdownify==0.7.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
mysqlclient==2.0.1
nodeenv==1.6.0
oauthlib==3.1.0
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.12.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==2.1.0
pylint==2.8.2
pylint-django==2.4.3
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
